Name standardise changes

+++ name_standardise.py
@@ -8,7 +8,7 @@ parser.add_argument('--mask_dir', type=str, default='out')
 parser.add_argument('--cutout_dir', type=str, default='out')
 parser.add_argument('--boundings_dir', type=str, default='out')
 parser.add_argument('--caption_dir', type=str, default='out')
-parser.add_argument('--prefix', type=str, default='img', help='prefix of images')
+parser.add_argument('--dry_run', type=bool, default=True, help='will not delete files if true')
 
 
 def main():
@@ -18,26 +18,42 @@ def main():
     cutout_dir = args.cutout_dir
     boundings_dir = args.boundings_dir
     caption_dir = args.caption_dir
+    dry_run = args.dry_run
 
     # Get all the images
     image_files = glob.glob(image_dir + '/*')
     print('Found {} image files'.format(len(image_files)))
+    if len(image_files) == 0:
+        print('No images found in {}'.format(image_dir))
+        return
 
     # Get all the masks
     mask_files = glob.glob(mask_dir + '/*')
     print('Found {} mask files'.format(len(mask_files)))
+    if len(mask_files) == 0:
+        print('No masks found in {}'.format(mask_dir))
+        return
 
     # Get all the cutouts
     cutout_files = glob.glob(cutout_dir + '/*')
     print('Found {} cutout files'.format(len(cutout_files)))
+    if len(cutout_files) == 0:
+        print('No cutouts found in {}'.format(cutout_dir))
+        return
 
     # Get all the boundings
     boundings_files = glob.glob(boundings_dir + '/*')
     print('Found {} boundings files'.format(len(boundings_files)))
+    if len(boundings_files) == 0:
+        print('No boundings found in {}'.format(boundings_dir))
+        return
 
     # Get all the captions
     caption_files = glob.glob(caption_dir + '/*')
     print('Found {} caption files'.format(len(caption_files)))
+    if len(caption_files) == 0:
+        print('No captions found in {}'.format(caption_dir))
+        return
 
     # Iterate through all the images
     for idx, image_file in enumerate(image_files):
@@ -62,14 +78,24 @@ def main():
         print('Found caption file {}'.format(len(caption_files)))
 
         # Rename the files
-        os.rename(image_file, image_file.replace(name, 'img_' + str(idx)))
-        os.rename(boundings_file, boundings_file.replace(name, 'bnd_' + str(idx)))
-        for mask_file in mask_files:
-            os.rename(mask_file, mask_file.replace('m_' + name, 'msk_' + str(idx)))
-        for cutout_file in cutout_files:
-            os.rename(cutout_file, cutout_file.replace('c_' + name, 'cut_' + str(idx)))
-        for caption_file in caption_files:
-            os.rename(caption_file, caption_file.replace('t_' + name, 'cap_' + str(idx)))
+        if dry_run:
+            print('Renaming {} to img_{}'.format(image_file, idx))
+            print('Renaming {} to bnd_{}'.format(boundings_file, idx))
+            for mask_file in mask_files:
+                print('Renaming {} to msk_{}'.format(mask_file, idx))
+            for cutout_file in cutout_files:
+                print('Renaming {} to cut_{}'.format(cutout_file, idx))
+            for caption_file in caption_files:
+                print('Renaming {} to cap_{}'.format(caption_file, idx))
+        else:
+            os.rename(image_file, image_file.replace(name, 'img_' + str(idx)))
+            os.rename(boundings_file, boundings_file.replace(name, 'bnd_' + str(idx)))
+            for mask_file in mask_files:
+                os.rename(mask_file, mask_file.replace('m_' + name, 'msk_' + str(idx)))
+            for cutout_file in cutout_files:
+                os.rename(cutout_file, cutout_file.replace('c_' + name, 'cut_' + str(idx)))
+            for caption_file in caption_files:
+                os.rename(caption_file, caption_file.replace('t_' + name, 'cap_' + str(idx)))
 
 
 if __name__ == '__main__':
